# background/docs/openapi3.json
{
  "components": {
    "schemas": {
      "types.Column": {
        "properties": {
          "exp": {
            "description": "expressions, which default to = when the value is null, have =, !=, \u003e, \u003e=, \u003c, \u003c=, like",
            "type": "string"
          },
          "logic": {
            "description": "logical type, default value is \"and\", support \u0026, and, ||, or",
            "type": "string"
          },
          "name": {
            "description": "column name",
            "type": "string"
          },
          "value": {
            "description": "column value"
          }
        },
        "type": "object"
      },
...
      "types.CreateUsersReply": {
        "properties": {
          "code": {
            "description": "return code",
            "type": "integer"
          },
          "data": {
            "description": "return data",
            "properties": {
              "id": {
                "description": "id",
                "type": "integer"
              }
            },
            "type": "object"
          },
          "msg": {
            "description": "return information description",
            "type": "string"
          }
        },
        "type": "object"
      },
      "types.CreateUsersRequest": {
        "properties": {
          "nickName": {
            "type": "string"
          },
          "password": {
            "type": "string"
          },
          "status": {
            "type": "string"
          },
          "userEmail": {
            "type": "string"
          },
          "userGender": {
            "type": "string"
          },
          "userName": {
            "type": "string"
          },
          "userPhone": {
            "type": "string"
          }
        },
        "type": "object"
      },
...
      "types.DeleteUsersByIDReply": {
        "properties": {
          "code": {
            "description": "return code",
            "type": "integer"
          },
          "data": {
            "description": "return data",
            "type": "object"
          },
          "msg": {
            "description": "return information description",
            "type": "string"
          }
        },
        "type": "object"
      },
...
      "types.GetUsersByIDReply": {
        "properties": {
          "code": {
            "description": "return code",
            "type": "integer"
          },
          "data": {
            "description": "return data",
            "properties": {
              "users": {
                "$ref": "#/components/schemas/types.UsersObjDetail"
              }
            },
            "type": "object"
          },
          "msg": {
            "description": "return information description",
            "type": "string"
          }
        },
        "type": "object"
      },
...
      "types.ListUserssReply": {
        "properties": {
          "code": {
            "description": "return code",
            "type": "integer"
          },
          "data": {
            "description": "return data",
            "properties": {
              "userss": {
                "items": {
                  "$ref": "#/components/schemas/types.UsersObjDetail"
                },
                "type": "array"
              }
            },
            "type": "object"
          },
          "msg": {
            "description": "return information description",
            "type": "string"
          }
        },
        "type": "object"
      },
...
      "types.Params": {
        "properties": {
          "columns": {
            "description": "query conditions",
            "items": {
              "$ref": "#/components/schemas/types.Column"
            },
            "type": "array"
          },
          "limit": {
            "description": "lines per page",
            "type": "integer"
          },
          "page": {
            "description": "page number, starting from page 0",
            "type": "integer"
          },
          "sort": {
            "description": "sorted fields, multi-column sorting separated by commas",
            "type": "string"
          }
        },
        "type": "object"
      },
...
      "types.UpdateUsersByIDReply": {
        "properties": {
          "code": {
            "description": "return code",
            "type": "integer"
          },
          "data": {
            "description": "return data",
            "type": "object"
          },
          "msg": {
            "description": "return information description",
            "type": "string"
          }
        },
        "type": "object"
      },
      "types.UpdateUsersByIDRequest": {
        "properties": {
          "id": {
            "description": "uint64 id",
            "type": "integer"
          },
          "nickName": {
            "type": "string"
          },
          "password": {
            "type": "string"
          },
          "status": {
            "type": "string"
          },
          "userEmail": {
            "type": "string"
          },
          "userGender": {
            "type": "string"
          },
          "userName": {
            "type": "string"
          },
          "userPhone": {
            "type": "string"
          }
        },
        "type": "object"
      },
...
      "types.UsersObjDetail": {
        "properties": {
          "createdAt": {
            "type": "string"
          },
          "id": {
            "description": "convert to uint64 id",
            "type": "integer"
          },
          "nickName": {
            "type": "string"
          },
          "password": {
            "type": "string"
          },
          "status": {
            "type": "string"
          },
          "updatedAt": {
            "type": "string"
          },
          "userEmail": {
            "type": "string"
          },
          "userGender": {
            "type": "string"
          },
          "userName": {
            "type": "string"
          },
          "userPhone": {
            "type": "string"
          }
        },
        "type": "object"
      }
    },
    "securitySchemes": {
      "BearerAuth": {
        "description": "Type Bearer your-jwt-token to Value",
        "in": "header",
        "name": "Authorization",
        "type": "apiKey"
      }
    }
  },
  "info": {
    "contact": {},
    "description": "http server api docs",
    "title": "godemo api docs",
    "version": "v1.0.0"
  },
  "openapi": "3.0.3",
  "paths": {
...
    "/api/v1/users": {
      "post": {
        "description": "Creates a new users entity using the provided data in the request body.",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/types.CreateUsersRequest"
              }
            }
          },
          "description": "users information",
          "required": true,
          "x-originalParamName": "data"
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/types.CreateUsersReply"
                }
              }
            },
            "description": "OK"
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "summary": "Create a new users",
        "tags": [
          "users"
        ]
      }
    },
    "/api/v1/users/list": {
      "post": {
        "description": "Returns a paginated list of users based on query filters, including page number and size.",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/types.Params"
              }
            }
          },
          "description": "query parameters",
          "required": true,
          "x-originalParamName": "data"
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/types.ListUserssReply"
                }
              }
            },
            "description": "OK"
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "summary": "Get a paginated list of userss by custom conditions",
        "tags": [
          "users"
        ]
      }
    },
    "/api/v1/users/{id}": {
      "delete": {
        "description": "Deletes a existing users identified by the given id in the path.",
        "parameters": [
          {
            "description": "id",
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/types.DeleteUsersByIDReply"
                }
              }
            },
            "description": "OK"
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "summary": "Delete a users by id",
        "tags": [
          "users"
        ]
      },
      "get": {
        "description": "Gets detailed information of a users specified by the given id in the path.",
        "parameters": [
          {
            "description": "id",
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/types.GetUsersByIDReply"
                }
              }
            },
            "description": "OK"
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "summary": "Get a users by id",
        "tags": [
          "users"
        ]
      },
      "put": {
        "description": "Updates the specified users by given id in the path, support partial update.",
        "parameters": [
          {
            "description": "id",
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/types.UpdateUsersByIDRequest"
              }
            }
          },
          "description": "users information",
          "required": true,
          "x-originalParamName": "data"
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/types.UpdateUsersByIDReply"
                }
              }
            },
            "description": "OK"
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "summary": "Update a users by id",
        "tags": [
          "users"
        ]
      }
    }
  },
  "servers": [
    {
      "url": "http://localhost:8080/"
    },
    {
      "url": "https://localhost:8080/"
    }
  ]
}